refactor: consolidate eval utility functions
